# module/JepRiaShowcase/App/service-rest/src/api-spec/goods/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "JepriaShowcase Goods REST API"
  },
  "servers": [
    {
      "url": "/{appContextPath}/{apiEndpoint}"
    }
  ],
  "tags": [
    {
      "name": "goods: entity",
      "description": "Entity operations (CRUD)"
    },
    {
      "name": "goods: search",
      "description": "Goods search operations"
    },
    {
      "name": "dict",
      "description": "Dictionaries"
    },
    {
      "name": "other",
      "description": "Other operations"
    }
  ],
  "paths": {
    "/goods/{recordId}": {
      "get": {
        "tags": [
          "goods: entity"
        ],
        "summary": "Get record by ID",
        "description": "",
        "operationId": "getRecordById-goods",
        "parameters": [
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/GoodsDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "goods: entity"
        ],
        "summary": "Update record",
        "description": "",
        "operationId": "update",
        "parameters": [
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json;charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/GoodsUpdateDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "goods: entity"
        ],
        "summary": "Delete record by ID",
        "description": "",
        "operationId": "deleteRecordById-goods",
        "parameters": [
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/feature": {
      "post": {
        "tags": [
          "gods: entity"
        ],
        "summary": "Create record",
        "description": "",
        "operationId": "create-goods",
        "requestBody": {
          "content": {
            "application/json;charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/GoodsCreateDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/goods/search/{searchId}/resultset-size": {
      "get": {
        "tags": [
          "goods: search"
        ],
        "summary": "Get size of a resultset for a search request",
        "description": "",
        "operationId": "getSearchResultsetSize",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Cache-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/goods/search/{searchId}/resultset/paged-by-{pageSize}/{page}": {
      "get": {
        "tags": [
          "goods: search"
        ],
        "summary": "Get paged resultset for a search request",
        "description": "",
        "operationId": "getResultsetPaged",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "pattern": "\\d+"
            }
          },
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Cache-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoodsDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/goods/option/feature-operator": {
      "get": {
        "tags": [
          "dict"
        ],
        "summary": "Get feature operator options",
        "description": "",
        "operationId": "getFeatureOperator",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OptionDtoInteger"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/feature/{featureId}/set-feature-responsible": {
      "post": {
        "tags": [
          "other"
        ],
        "summary": "Set responsible operator for the feature",
        "description": "",
        "operationId": "setFeatureResponsible",
        "parameters": [
          {
            "name": "featureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "responsibleId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/goods/option/goods-segment": {
      "get": {
        "tags": [
          "dict"
        ],
        "summary": "Get Goods Segment options",
        "description": "",
        "operationId": "getGoodsSegment",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OptionDtoString"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/goods/option/motivation-type": {
      "get": {
        "tags": [
          "dict"
        ],
        "summary": "Get MotivationType options",
        "description": "",
        "operationId": "getMotivationType",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OptionDtoString"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/goods/option/unit": {
      "get": {
        "tags": [
          "dict"
        ],
        "summary": "Get Unit options",
        "description": "",
        "operationId": "getUnit",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OptionDtoString"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/goods/option/goods-type": {
      "get": {
        "tags": [
          "dict"
        ],
        "summary": "Get Goods Type options",
        "description": "",
        "operationId": "getGoodsType",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OptionDtoString"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/goods/search": {
      "post": {
        "tags": [
          "goods: search"
        ],
        "summary": "Post search request",
        "description": "",
        "operationId": "postSearch",
        "parameters": [
          {
            "name": "Extended-Response",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Cache-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json;charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequestDtoGoodsSearchDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/goods/search/{searchId}/resultset": {
      "get": {
        "tags": [
          "goods: search"
        ],
        "summary": "Get resultset for a search request",
        "description": "",
        "operationId": "getResultset",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Cache-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoodsDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/goods/search/{searchId}": {
      "get": {
        "tags": [
          "goods: search"
        ],
        "summary": "Get search request by ID",
        "description": "",
        "operationId": "getSearchRequest",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/SearchRequestDtoGoodsSearchDto"
                }
              }
            }
          }
        }
      }
    },
    "/feature/{featureId}/feature-process": {
      "get": {
        "tags": [
          "feature-process: entity"
        ],
        "summary": "List all records for the single feature by ID",
        "description": "",
        "operationId": "findFeatureProcess",
        "parameters": [
          {
            "name": "featureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeatureProcessDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "feature-process: entity"
        ],
        "summary": "Create record",
        "description": "",
        "operationId": "create-feature-process",
        "parameters": [
          {
            "name": "featureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json;charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/FeatureProcessCreateDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/feature/{featureId}/feature-process/{recordId}": {
      "get": {
        "tags": [
          "feature-process: entity"
        ],
        "summary": "Get record by ID",
        "description": "",
        "operationId": "getRecordById-feature-process",
        "parameters": [
          {
            "name": "featureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureProcessDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "feature-process: entity"
        ],
        "summary": "Delete record by ID",
        "description": "",
        "operationId": "deleteRecordById-feature-process",
        "parameters": [
          {
            "name": "featureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GoodsUpdateDto": {
        "type": "object",
        "properties": {
          "supplierId": {
            "type": "integer",
            "format": "int32"
          },
          "goodsName": {
            "type": "string"
          },
          "goodsTypeCode": {
            "type": "string"
          },
          "unitCode": {
            "type": "string"
          },
          "purchasingPrice": {
            "type": "integer",
            "format": "int64"
          },
          "motivationTypeCode": {
            "type": "string"
          },
          "goodsPhotoMimeType": {
            "type": "string"
          },
          "goodsPhotoExtension": {
            "type": "string"
          },
          "goodsPortfolioMimeType": {
            "type": "string"
          },
          "goodsPortfolioExtension": {
            "type": "string"
          }
        }
      },
      "GoodsCreateDto": {
        "type": "object",
        "properties": {
          "supplierId": {
            "type": "integer",
            "format": "int32"
          },
          "goodsName": {
            "type": "string"
          },
          "goodsTypeCode": {
            "type": "string"
          },
          "unitCode": {
            "type": "string"
          },
          "purchasingPrice": {
            "type": "integer",
            "format": "int64"
          },
          "motivationTypeCode": {
            "type": "string"
          },
          "goodsPhotoMimeType": {
            "type": "string"
          },
          "goodsPhotoExtension": {
            "type": "string"
          },
          "goodsPortfolioMimeType": {
            "type": "string"
          },
          "goodsPortfolioExtension": {
            "type": "string"
          }
        }
      },
      "GoodsDto": {
        "type": "object",
        "properties": {
          "goodsId": {
            "type": "integer",
            "format": "int32"
          },
          "supplierId": {
            "type": "integer",
            "format": "int32"
          },
          "goodsName": {
            "type": "string"
          },
          "goodsTypeCode": {
            "$ref": "#/components/schemas/OptionDtoString"
          },
          "unitCode": {
            "$ref": "#/components/schemas/OptionDtoString"
          },
          "purchasingPrice": {
            "type": "integer",
            "format": "int64"
          },
          "motivationTypeCode": {
            "$ref": "#/components/schemas/OptionDtoString"
          },
          "goodsPhotoMimeType": {
            "type": "string"
          },
          "goodsPhotoExtension": {
            "type": "string"
          },
          "goodsPortfolioMimeType": {
            "type": "string"
          },
          "goodsPortfolioExtension": {
            "type": "string"
          },
          "featureStatus": {
            "$ref": "#/components/schemas/OptionDtoString"
          },
          "dateIns": {
            "type": "string",
            "format": "date-time"
          },
          "author": {
            "$ref": "#/components/schemas/OptionDtoInteger"
          },
          "responsible": {
            "$ref": "#/components/schemas/OptionDtoInteger"
          }
        }
      },
      "OptionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "object"
          }
        }
      },
      "OptionDtoInteger": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "OptionDtoString": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "ColumnSortConfigurationDto": {
        "type": "object",
        "properties": {
          "columnName": {
            "type": "string"
          },
          "sortOrder": {
            "type": "string"
          }
        }
      },
      "GoodsSearchDto": {
        "type": "object",
        "properties": {
          "featureId": {
            "type": "integer",
            "format": "int32"
          },
          "featureNameTemplate": {
            "type": "string"
          },
          "featureNameEnTemplate": {
            "type": "string"
          },
          "statusCodeList": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "authorId": {
            "type": "integer",
            "format": "int32"
          },
          "responsibleId": {
            "type": "integer",
            "format": "int32"
          },
          "dateInsFrom": {
            "type": "string",
            "format": "date-time"
          },
          "dateInsTo": {
            "type": "string",
            "format": "date-time"
          },
          "maxRowCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SearchRequestDto": {
        "type": "object",
        "properties": {
          "template": {
            "type": "object"
          },
          "listSortConfiguration": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnSortConfigurationDto"
            }
          }
        }
      },
      "SearchRequestDtoGoodsSearchDto": {
        "type": "object",
        "properties": {
          "template": {
            "$ref": "#/components/schemas/GoodsSearchDto"
          },
          "listSortConfiguration": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnSortConfigurationDto"
            }
          }
        }
      },
      "FeatureProcessCreateDto": {
        "type": "object",
        "properties": {
          "featureStatusCode": {
            "type": "string"
          }
        }
      },
      "FeatureProcessDto": {
        "type": "object",
        "properties": {
          "featureProcessId": {
            "type": "integer",
            "format": "int32"
          },
          "featureId": {
            "type": "integer",
            "format": "int32"
          },
          "featureStatusCode": {
            "type": "string"
          },
          "featureStatusName": {
            "type": "string"
          },
          "dateIns": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}